Fix failing e2e tests after renaming AutomatedChecks* to Results*
