+++ src/tests/electron/common/view-controllers/automated-checks-view-controller.ts
@@ -28,23 +28,4 @@ export class AutomatedChecksViewController extends ViewController {
         await this.waitForSelector(selector);
         await this.client.click(selector);
     }
-
-    public async setToggleState(toggleSelector: string, newState: boolean): Promise<void> {
-        await this.waitForSelector(toggleSelector);
-        const oldState = await this.client.getAttribute(toggleSelector, 'aria-checked');
-
-        const oldStateBool = oldState.toLowerCase() === 'true';
-        if (oldStateBool !== newState) {
-            await this.click(toggleSelector);
-            await this.expectToggleState(toggleSelector, newState);
-        }
-    }
-
-    public async expectToggleState(toggleSelector: string, expectedState: boolean): Promise<void> {
-        const toggleInStateSelector = expectedState
-            ? settingsPanelSelectors.enabledToggle(toggleSelector)
-            : settingsPanelSelectors.disabledToggle(toggleSelector);
-
-        await this.waitForSelector(toggleInStateSelector);
-    }
 }
