+++ src/tests/electron/tests/unified-settings-panel.test.ts
@@ -4,13 +4,13 @@ import * as path from 'path';
 import { createApplication } from 'tests/electron/common/create-application';
 import { scanForAccessibilityIssuesInAllModes } from 'tests/electron/common/scan-for-accessibility-issues';
 import { AppController } from 'tests/electron/common/view-controllers/app-controller';
-import { AutomatedChecksViewController } from 'tests/electron/common/view-controllers/automated-checks-view-controller';
+import { ResultsViewController } from 'tests/electron/common/view-controllers/results-view-controller';
 import { CommonSelectors } from 'tests/end-to-end/common/element-identifiers/common-selectors';
 import { settingsPanelSelectors } from 'tests/end-to-end/common/element-identifiers/details-view-selectors';
 import { commonAdbConfigs, setupMockAdb } from 'tests/miscellaneous/mock-adb/setup-mock-adb';
 describe('AutomatedChecksView -> Settings Panel', () => {
     let app: AppController;
-    let automatedChecksView: AutomatedChecksViewController;
+    let resultsView: ResultsViewController;
 
     beforeEach(async () => {
         await setupMockAdb(
@@ -19,9 +19,9 @@ describe('AutomatedChecksView -> Settings Panel', () => {
             'beforeEach',
         );
         app = await createApplication({ suppressFirstTimeDialog: true });
-        automatedChecksView = await app.openResultsView();
-        await automatedChecksView.waitForViewVisible();
-        await automatedChecksView.openSettingsPanel();
+        resultsView = await app.openResultsView();
+        await resultsView.waitForViewVisible();
+        await resultsView.openSettingsPanel();
     });
 
     afterEach(async () => {
@@ -32,24 +32,15 @@ describe('AutomatedChecksView -> Settings Panel', () => {
 
     describe('Telemetry toggle', () => {
         it('should be "off" as per our suppressFirstTimeDialog implementation of the initial state', async () => {
-            await automatedChecksView.expectToggleState(
-                settingsPanelSelectors.telemetryStateToggle,
-                false,
-            );
+            await resultsView.expectToggleState(settingsPanelSelectors.telemetryStateToggle, false);
         });
 
         it('should reflect the state applied via the insightsUserConfiguration controller', async () => {
             await app.setTelemetryState(true);
-            await automatedChecksView.expectToggleState(
-                settingsPanelSelectors.telemetryStateToggle,
-                true,
-            );
+            await resultsView.expectToggleState(settingsPanelSelectors.telemetryStateToggle, true);
 
             await app.setTelemetryState(false);
-            await automatedChecksView.expectToggleState(
-                settingsPanelSelectors.telemetryStateToggle,
-                false,
-            );
+            await resultsView.expectToggleState(settingsPanelSelectors.telemetryStateToggle, false);
         });
     });
 
@@ -59,39 +50,31 @@ describe('AutomatedChecksView -> Settings Panel', () => {
         it.skip('should default to system setting or non-high-contrast mode', async () => {
             const highContrastModeEnabled = false;
 
-            await automatedChecksView.expectToggleState(
+            await resultsView.expectToggleState(
                 settingsPanelSelectors.highContrastModeToggle,
                 highContrastModeEnabled,
             );
         });
 
         it('should apply the appropriate CSS style when High Contrast Mode setting is toggled', async () => {
-            await automatedChecksView.setToggleState(
-                settingsPanelSelectors.highContrastModeToggle,
-                true,
-            );
+            await resultsView.setToggleState(settingsPanelSelectors.highContrastModeToggle, true);
 
-            await automatedChecksView.waitForSelector(CommonSelectors.highContrastThemeSelector);
+            await resultsView.waitForSelector(CommonSelectors.highContrastThemeSelector);
 
-            await automatedChecksView.setToggleState(
-                settingsPanelSelectors.highContrastModeToggle,
-                false,
-            );
+            await resultsView.setToggleState(settingsPanelSelectors.highContrastModeToggle, false);
 
-            await automatedChecksView.waitForSelectorToDisappear(
-                CommonSelectors.highContrastThemeSelector,
-            );
+            await resultsView.waitForSelectorToDisappear(CommonSelectors.highContrastThemeSelector);
         });
 
         it('should reflect the state applied via the insightsUserConfiguration controller', async () => {
             await app.setHighContrastMode(true);
-            await automatedChecksView.expectToggleState(
+            await resultsView.expectToggleState(
                 settingsPanelSelectors.highContrastModeToggle,
                 true,
             );
 
             await app.setHighContrastMode(false);
-            await automatedChecksView.expectToggleState(
+            await resultsView.expectToggleState(
                 settingsPanelSelectors.highContrastModeToggle,
                 false,
             );
